+++ src/rules/enforce-relation-types.test.ts
@@ -186,6 +186,58 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
                 others: Relation<Other[]>;
             }`,
         },
+        {
+            name: 'should allow undefined via optional marker one-to-many relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToMany(() => Other, (other) => other.entity)
+                others?: Other[];
+            }`,
+        },
+        {
+            name: 'should allow undefined via union with undefined one-to-many relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToMany(() => Other, (other) => other.entity)
+                others: Other[] | undefined;
+            }`,
+        },
+        {
+            name: 'should allow undefined via optional marker one-to-one relations',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity)
+                @JoinColumn()
+                other?: Other | null;
+            }`,
+        },
+        {
+            name: 'should allow undefined via union with undefined one-to-one relations',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity)
+                @JoinColumn()
+                other: Other | undefined | null;
+            }`,
+        },
+        {
+            name: 'should allow omission of undefined for lazy relations',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity, { nullable: false })
+                @JoinColumn()
+                other: Promise<Other>;
+            }`,
+        },
+        {
+            name: 'should allow omission of undefined for eager relations',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity, { nullable: false, eager: true })
+                @JoinColumn()
+                other: Other;
+            }`,
+        },
     ],
     invalid: [
         {
@@ -853,6 +905,142 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
                 @ManyToMany(() => Other)
                 @JoinTable()
                 others: Other[];
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on undefined one-to-many relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToMany(() => Other, (other) => other.entity)
+                others: Other[];
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_specify_undefined_always',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @OneToMany(() => Other, (other) => other.entity)
+                others: Other[] | undefined;
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on undefined many-to-one relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @ManyToOne(() => Other, { nullable: false })
+                other: Other;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_specify_undefined_always',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @ManyToOne(() => Other, { nullable: false })
+                other: Other | undefined;
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on undefined one-to-one relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity)
+                @JoinColumn()
+                other: Other | null;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_specify_undefined_always',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity)
+                @JoinColumn()
+                other: Other | null | undefined;
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on undefined eager one-to-many relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToMany(() => Other, (other) => other.entity, { eager: true })
+                others: Other[] | undefined;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_specify_undefined_always',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @OneToMany(() => Other, (other) => other.entity, { eager: true })
+                others: Other[];
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on undefined eager many-to-one relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @ManyToOne(() => Other, { nullable: false, eager: true })
+                other: Other | undefined;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_specify_undefined_always',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @ManyToOne(() => Other, { nullable: false, eager: true })
+                other: Other;
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on undefined eager one-to-one relations ',
+            options: [{ specifyUndefined: 'always' }],
+            code: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity, { eager: true })
+                @JoinColumn()
+                other: Other | null | undefined;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_specify_undefined_always',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @OneToOne(() => Other, (other) => other.entity, { eager: true })
+                @JoinColumn()
+                other: Other | null;
             }`,
                         },
                     ],